requirements.txt
colorama==0.4.5
discord.py==2.0.1
pypresence==4.2.1
requests==2.28.1
